+++ functions/src/callable/trigger-password-reset.ts
@@ -1,14 +1,29 @@
-import * as functions from 'firebase-functions';
-import {auth} from 'firebase-admin';
 import {SHARED_CONFIG} from 'definitions';
+import {auth} from 'firebase-admin';
+import * as functions from 'firebase-functions';
 import {EmailService} from '../services/email/email.service';
 import {hasRole} from '../utils/auth';
+import {schemaValidation} from '../utils/schema-validation';
 
 export const triggerPasswordReset = functions
   .region(SHARED_CONFIG.cloudRegion)
   .https
   .onCall(async (data, context) => {
     hasRole(context, 'admin');
+    schemaValidation(
+      data,
+      {
+        additionalProperties: false,
+        properties: {
+          email: {type: 'string'},
+          url: {type: 'string'}
+        },
+        required: [
+          'email',
+          'url'
+        ]
+      }
+    );
 
     let link;
 
@@ -24,7 +39,7 @@ export const triggerPasswordReset = functions
         to: data.email
       });
     } catch (e) {
-      console.error(e);
+      functions.logger.error(e);
       throw new functions.https.HttpsError('internal', e.toString());
     }