+++ CHANGELOG.md
@@ -1,3 +1,97 @@
+## [5.41.3](https://github.com/jaspero/jms/compare/v5.41.2...v5.41.3) (2022-10-04)
+
+
+### Bug Fixes
+
+* functions workflow ([1b932e2](https://github.com/jaspero/jms/commit/1b932e2d32e5be7342d0b484e13669cc65a5e33d))
+
+## [5.41.2](https://github.com/jaspero/jms/compare/v5.41.1...v5.41.2) (2022-10-04)
+
+
+### Bug Fixes
+
+* functions workflow ([662e5cc](https://github.com/jaspero/jms/commit/662e5ccd1c7ec20704c77b438ea31c14069047c1))
+
+## [5.41.1](https://github.com/jaspero/jms/compare/v5.41.0...v5.41.1) (2022-10-04)
+
+
+### Bug Fixes
+
+* functions workflow ([6bbe80c](https://github.com/jaspero/jms/commit/6bbe80cdbdf025ba2f2ee55283a669c724ba2be1))
+
+# [5.41.0](https://github.com/jaspero/jms/compare/v5.40.0...v5.41.0) (2022-10-04)
+
+
+### Features
+
+* extended functions workflow with options for running all functions ([75f4156](https://github.com/jaspero/jms/commit/75f4156e292a0178d7b347abfb079f15092e3eda))
+
+# [5.40.0](https://github.com/jaspero/jms/compare/v5.39.7...v5.40.0) (2022-10-04)
+
+
+### Features
+
+* **functions:** replaced all console logs with functions.logs closes [#411](https://github.com/jaspero/jms/issues/411) ([db0f183](https://github.com/jaspero/jms/commit/db0f183b193a5a19e06d8ac39adf5e13a112b409))
+
+## [5.39.7](https://github.com/jaspero/jms/compare/v5.39.6...v5.39.7) (2022-10-01)
+
+
+### Bug Fixes
+
+* **cms:** keys in email templates properly represented with [[ ]] ([94f902f](https://github.com/jaspero/jms/commit/94f902f1b4bcfb5cb0079a19f76fcc55804cedc7))
+
+## [5.39.6](https://github.com/jaspero/jms/compare/v5.39.5...v5.39.6) (2022-09-22)
+
+
+### Bug Fixes
+
+* **scripts:** backfill search not creating documents ([9cf1b54](https://github.com/jaspero/jms/commit/9cf1b548973cd2d969e3de6d3229ed9ce6944b09))
+
+## [5.39.5](https://github.com/jaspero/jms/compare/v5.39.4...v5.39.5) (2022-09-22)
+
+
+### Bug Fixes
+
+* **jms:** message format ([be1121b](https://github.com/jaspero/jms/commit/be1121bb935aa0366880cb4262ad6fb2d953041a))
+
+## [5.39.4](https://github.com/jaspero/jms/compare/v5.39.3...v5.39.4) (2022-09-20)
+
+
+### Bug Fixes
+
+* **cms:** don't show Dynamic Values on sample email if there aren't any fields ([dc553f9](https://github.com/jaspero/jms/commit/dc553f9a4d7527a3c87f5e944f852b70591834b8))
+* **cms:** missing role on permissions ([93779a9](https://github.com/jaspero/jms/commit/93779a9bcc79e3f2c03a1129417d174bd6dd625d))
+* **cms:** showing empty dynamic values on email component ([caaac37](https://github.com/jaspero/jms/commit/caaac37338423459c6655f9c8f0d73a38da7fd8e))
+* **functions:** sample-email missing ([f2501ea](https://github.com/jaspero/jms/commit/f2501eaef771e6e940bd0d48ea4cc644447228ca))
+
+## [5.39.3](https://github.com/jaspero/jms/compare/v5.39.2...v5.39.3) (2022-08-28)
+
+
+### Bug Fixes
+
+* **definitions:** missing properties on user ([7ac75f2](https://github.com/jaspero/jms/commit/7ac75f214ab055b299955cb7c890b177c37f8783))
+
+## [5.39.2](https://github.com/jaspero/jms/compare/v5.39.1...v5.39.2) (2022-08-28)
+
+
+### Bug Fixes
+
+* jsx values not applied properly if they have spaces ([1ca1c0a](https://github.com/jaspero/jms/commit/1ca1c0a742ea57d359969d181e01570c7fd996d6))
+
+## [5.39.1](https://github.com/jaspero/jms/compare/v5.39.0...v5.39.1) (2022-08-28)
+
+
+### Bug Fixes
+
+* **cms:** showing `id` when editTitleKey is undefined ([82e9561](https://github.com/jaspero/jms/commit/82e9561452f4dcd60911d77e9d397abd0c7af22b))
+
+# [5.39.0](https://github.com/jaspero/jms/compare/v5.38.0...v5.39.0) (2022-08-27)
+
+
+### Features
+
+* added createdBy to users; connected search in ref; ([ba75348](https://github.com/jaspero/jms/commit/ba7534804667af241203d8407323d63d9fb05022))
+
 # [5.38.0](https://github.com/jaspero/jms/compare/v5.37.1...v5.38.0) (2022-08-05)